move TestExecutionMetaData route registration and handler to grid.ts
