+++ src/app/routers/grid.ts
@@ -14,6 +14,7 @@ import { DeviceFarmManager } from '../../device-managers';
 import Container from 'typedi';
 import { IPluginArgs } from '../../interfaces/IPluginArgs';
 import { IDevice } from '../../interfaces/IDevice';
+import { saveTestExecutionMetaData } from '../../wdio-service/wdio-service';
 
 const SERVER_UP_TIME = new Date().toISOString();
 
@@ -226,6 +227,19 @@ async function getDevicesFromDeviceManager() {
   return devices;
 }
 
+ async function handleTestExecutionMetaData(req: Request, res: Response) {
+  try {
+    await saveTestExecutionMetaData(req);
+    log.info("Saved Test Execution Meta Data.")
+    res.status(200).send('{"message": "Saved Test Execution Meta Data"}');
+  } catch(e) {
+    const response = {message: `Failed to save Test Execution Meta Data. Error: ${e}`}
+    log.error(`Error while handling TestExecutionMetaData.`);
+    log.error(`Sending response - ${response}.`)
+    res.status(500).json(response);
+  }
+}
+
 function register(router: Router, pluginArgs: IPluginArgs) {
   router.get('/device', getDevices);
   router.get('/device/:platform', getDeviceByPlatform);
@@ -252,6 +266,9 @@ function register(router: Router, pluginArgs: IPluginArgs) {
       });
     },
   );
+
+  // test execution meta data
+  router.post('/handleTestExecutionMetaData', handleTestExecutionMetaData);
 }
 
 export default {
